+++ controllers/index.js
@@ -1,13 +1,11 @@
 const router = require("express").Router();
 
-const apiRoutes = require("./api");
 const homeRoutes = require("./home-routes");
-const dashboardRoutes = require("./dashboard");
 const userRoutes = require("./user-routes");
+const apiRoutes = require("./api");
 
 router.use("/", homeRoutes);
 router.use("/user", userRoutes);
 router.use("/api", apiRoutes);
-router.use("/my-dashboard", dashboardRoutes);
 
 module.exports = router;
