+++ controllers/user-routes.js
@@ -1,6 +1,25 @@
 const router = require("express").Router();
 const { User, Dish, Comment } = require("../models");
 
+//GET User by Id
+router.get("/:id", async (req, res) => {
+  try {
+    const userData = await User.findByPk(req.params.id, {
+      include: [{ model: Dish }],
+    });
+    const userPosts = userData.get({ plain: true });
+
+    if (!userData) {
+      res.status(404).json({ message: "No user found with that id!" });
+      return;
+    }
+    console.log(userPosts);
+    res.render("timeline", { userPosts });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
 router.get("/login", async (req, res) => {
   try {
     const userData = await User.findAll({
