+++ views/layouts/main.handlebars
@@ -9,9 +9,7 @@
     <title>Document</title>
 </head>
 <body>
-    {{> navbar}}
-    {{> newDish}}
-    {{> newDishModal}}
+    {{>navbar}}
     {{{body}}}
 <script src="/index.js"></script>
 </body>
