--- views/layouts/timeline.handlebars
@@ -1,3 +1,0 @@
-{{#data}}
-    {{> card dishName=dishName author=author description=description availability=availability price=price}}
-{{/data}}
